Crate rendy_shader

Shader compilation.

Structs

ShaderSet
A ShaderSet object represents a merged collection of ShaderStorage structures, which reflects merged information for all shaders in the set.

ShaderSetBuilder
Builder class which is used to begin the reflection and shader set construction process for a shader set. Provides all the functionality needed to build a shader set with provided shaders and then reflect appropriate gfx-hal and generic shader information.

ShaderStorage
Contains reflection and runtime nformation for a given compiled Shader Module.

SpecConstantSet
A set of Specialization constants for a certain shader set.

SpirvShader
Spir-V shader.

Enums

ShaderError
Error type returned by this module.

Traits

Shader
Interface to create shader modules from shaders. Implemented for static shaders via [compile_to_spirv!] macro.